Add first implementation
This initializes the Go project, setting-up the http server together
with its handlers.
Logic will be kept in services for reusability reasons, so `product`
package contains Go logic for resolving product use-cases. Services
must depend on nothing but the root (domain) package. Interfaces are
used to abstract dependencies.
Current implementation uses a mock of what could be a `PackRepository`,
the place from where the configuration can be received.
`cmd` package ties together all dependencies, also serving as the entry
point for the application.