Add configuration update
Updating packaging configuration is now exposed on the same path under
PUT method. If no configuration has been set by user, will fallback to
the default configuration.
Updated mock and add tests for configuration update.
Move computing algorithms to separate files. It's to be decided if the
algorithms should be moved to separate packages and be individually
tested.